refactor : 댓글 이미지 조회 수정(댓글 조회 시 병합), 유틸 함수 추가, DTO 검증 및 필드 순서 조정, 서비스 수정
1. 댓글 이미지 조회 수정(댓글 조회 시 병합)
 - getAllCommentList() 모든 댓글 리스트 조회 시 이미지 조회 병합
 - 쿼리 조정, join 으로 댓글 조회 시 해당 댓글의 이미지 리스트 조회
 - 기존 이미지 조회 중복 코드 제거

2. 유틸 함수 추가
 - UserUtil 소유자 검사, 인증 사용자 검증 메서드 추가

3. DTO 검증 및 필드 순서 조정
 - DTO NotNull Valid 검증 추가
 - 필드 순서 조정

4. 서비스 수정
 - 서비스 코드 중복 메서드 제거(댓글 검증)

+++ src/main/java/arraysort/project/board/app/comment/domain/CommentAddReqDTO.java
@@ -22,10 +22,10 @@ public class CommentAddReqDTO {
 
 	private Long parentId;
 
-	private Long depth;
+	private Integer depth;
 
 	// 최상위 댓글 ID, 댓글 depth 설정
-	public void setCommentInfo(Long topParentId, Long depth) {
+	public void setCommentInfo(Long topParentId, int depth) {
 		this.topParentId = topParentId;
 		this.depth = depth;
 	}
